+++ axum/src/json.rs
@@ -1,10 +1,10 @@
 use crate::{
-    extract::{rejection::*, take_body, FromRequest, RequestParts},
+    body::{self, BoxBody},
+    extract::{rejection::*, FromRequest, RequestParts},
     response::IntoResponse,
     BoxError,
 };
 use async_trait::async_trait;
-use bytes::Bytes;
 use http::{
     header::{self, HeaderValue},
     StatusCode,
@@ -12,10 +12,7 @@ use http::{
 use http_body::Full;
 use hyper::Response;
 use serde::{de::DeserializeOwned, Serialize};
-use std::{
-    convert::Infallible,
-    ops::{Deref, DerefMut},
-};
+use std::ops::{Deref, DerefMut};
 
 /// JSON Extractor / Response.
 ///
@@ -101,16 +98,10 @@ where
     type Rejection = JsonRejection;
 
     async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        use bytes::Buf;
-
         if json_content_type(req)? {
-            let body = take_body(req)?;
+            let bytes = bytes::Bytes::from_request(req).await?;
 
-            let buf = hyper::body::aggregate(body)
-                .await
-                .map_err(InvalidJsonBody::from_err)?;
-
-            let value = serde_json::from_reader(buf.reader()).map_err(InvalidJsonBody::from_err)?;
+            let value = serde_json::from_slice(&bytes).map_err(InvalidJsonBody::from_err)?;
 
             Ok(Json(value))
         } else {
@@ -122,7 +113,7 @@ where
 fn json_content_type<B>(req: &RequestParts<B>) -> Result<bool, HeadersAlreadyExtracted> {
     let content_type = if let Some(content_type) = req
         .headers()
-        .ok_or(HeadersAlreadyExtracted)?
+        .ok_or_else(HeadersAlreadyExtracted::default)?
         .get(header::CONTENT_TYPE)
     {
         content_type
@@ -172,25 +163,25 @@ impl<T> IntoResponse for Json<T>
 where
     T: Serialize,
 {
-    type Body = Full<Bytes>;
-    type BodyError = Infallible;
-
-    fn into_response(self) -> Response<Self::Body> {
+    fn into_response(self) -> Response<BoxBody> {
         let bytes = match serde_json::to_vec(&self.0) {
             Ok(res) => res,
             Err(err) => {
                 return Response::builder()
                     .status(StatusCode::INTERNAL_SERVER_ERROR)
-                    .header(header::CONTENT_TYPE, "text/plain")
-                    .body(Full::from(err.to_string()))
+                    .header(
+                        header::CONTENT_TYPE,
+                        HeaderValue::from_static(mime::TEXT_PLAIN_UTF_8.as_ref()),
+                    )
+                    .body(body::boxed(Full::from(err.to_string())))
                     .unwrap();
             }
         };
 
-        let mut res = Response::new(Full::from(bytes));
+        let mut res = Response::new(body::boxed(Full::from(bytes)));
         res.headers_mut().insert(
             header::CONTENT_TYPE,
-            HeaderValue::from_static("application/json"),
+            HeaderValue::from_static(mime::APPLICATION_JSON.as_ref()),
         );
         res
     }